Crate wasmer_compiler_cranelift

A WebAssembly Compiler implementation using Cranelift.

Cranelift is a fast IR generator created by Mozilla for usage in Firefox as a next JS compiler generator.

Compared to LLVM, Cranelift is a bit faster and made entirely in Rust.

Modules

address_map 🔒

compiler 🔒

Support for compiling with Cranelift.

config 🔒

debug 🔒

dwarf 🔒

func_environ 🔒

heap 🔒

Heaps to implement WebAssembly linear memories.

table 🔒

trampoline 🔒

translator 🔒

Tools for translating wasm function bytecode to Cranelift IR.

Structs

Cranelift

Global configuration options used to create an wasmer_engine::Engine and customize its behavior.

CraneliftCompiler

A compiler that compiles a WebAssembly module with Cranelift, translating the Wasm to Cranelift IR, optimizing it and then translating to assembly.

ModuleInfoVmctxInfo

ModuleInfo vmctx related info.

Enums

CraneliftOptLevel

Possible optimization levels for the Cranelift codegen backend.

ModuleInfoMemoryOffset

Memory definition offset in the VMContext structure.

Constants

VERSION

Version number of this crate.

Functions

make_trampoline_function_call

Create a trampoline for invoking a WebAssembly function.

Type Aliases

ValueLabelsRanges

Value ranges for functions.