Struct wasmer_compiler_cranelift::sink::RelocSink

pub(crate) struct RelocSink<'a> {
    module: &'a ModuleInfo,
    local_func_index: LocalFunctionIndex,
    pub func_relocs: Vec<Relocation>,
    pub probestack_trampoline_relocation_target: SectionIndex,
}

Implementation of a relocation sink that just saves all the information for later

Fields
module: &'a ModuleInfolocal_func_index: LocalFunctionIndex

Current function index.

func_relocs: Vec<Relocation>

Relocations recorded for the function.

probestack_trampoline_relocation_target: SectionIndex

The section where the probestack trampoline call is located

Implementations

impl<'a> RelocSink<'a>

pub fn new(
module: &'a ModuleInfo,
func_index: FunctionIndex,
probestack_trampoline_relocation_target: SectionIndex
) -> Self

Return a new RelocSink instance.