wasmer_compiler_cranelift

Struct Cranelift

pub struct Cranelift {
    enable_nan_canonicalization: bool,
    enable_verifier: bool,
    pub(crate) enable_perfmap: bool,
    enable_pic: bool,
    opt_level: CraneliftOptLevel,
    pub num_threads: NonZero<usize>,
    pub(crate) middlewares: Vec<Arc<dyn ModuleMiddleware>>,
    pub(crate) callbacks: Option<CraneliftCallbacks>,
}

Global configuration options used to create an wasmer_engine::Engine and customize its behavior.

This structure exposes a builder-like interface and is primarily consumed by wasmer_engine::Engine::new.

Fields
enable_nan_canonicalization: boolenable_verifier: boolenable_perfmap: boolenable_pic: boolopt_level: CraneliftOptLevelnum_threads: NonZero<usize>

The number of threads to use for compilation.

middlewares: Vec<Arc<dyn ModuleMiddleware>>

The middleware chain.

callbacks: Option<CraneliftCallbacks>
Implementations

impl Cranelift

pub fn new() -> Self

Creates a new configuration object with the default configuration specified.

pub fn canonicalize_nans(&mut self, enable: bool) -> &mut Self

Enable NaN canonicalization.

NaN canonicalization is useful when trying to run WebAssembly deterministically across different architectures.

pub fn num_threads(&mut self, num_threads: NonZero<usize>) -> &mut Self

Set the number of threads to use for compilation.

pub fn opt_level(&mut self, opt_level: CraneliftOptLevel) -> &mut Self

The optimization levels when optimizing the IR.

pub fn isa(&self, target: &Target) -> CodegenResult<Arc<dyn TargetIsa>>

Generates the ISA for the provided target

pub fn flags(&self) -> Flags

Generates the flags for the compiler

pub fn callbacks(&mut self, callbacks: Option<CraneliftCallbacks>) -> &mut Self

Callbacks that will triggered in the different compilation phases in Cranelift.