Struct wasmer_compiler_cranelift::trampoline::Context

pub struct Context {
    pub func: Function,
    pub cfg: ControlFlowGraph,
    pub domtree: DominatorTree,
    pub loop_analysis: LoopAnalysis,
    pub(crate) compiled_code: Option<CompiledCodeBase<Final>>,
    pub want_disasm: bool,
}

Persistent data structures and compilation pipeline.

Fields
func: Function

The function weโre compiling.

cfg: ControlFlowGraph

The control flow graph of func.

domtree: DominatorTree

Dominator tree for func.

loop_analysis: LoopAnalysis

Loop analysis of func.

compiled_code: Option<CompiledCodeBase<Final>>want_disasm: bool

Flag: do we want a disassembly with the CompiledCode?

Implementations

impl Context

pub fn new() -> Context

Allocate a new compilation context.

The returned instance should be reused for compiling multiple functions in order to avoid needless allocator thrashing.

pub fn for_function(func: Function) -> Context

Allocate a new compilation context with an existing Function.

The returned instance should be reused for compiling multiple functions in order to avoid needless allocator thrashing.

pub fn clear(&mut self)

Clear all data structures in this context.

pub fn compiled_code(&self) -> Option<&CompiledCodeBase<Final>>

Returns the compilation result for this function, available after any compile function has been called.

pub fn set_disasm(&mut self, val: bool)

Set the flag to request a disassembly when compiling with a MachBackend backend.

pub fn compile_and_emit( &mut self, isa: &dyn TargetIsa, mem: &mut Vec<u8, Global> ) -> Result<&CompiledCodeBase<Final>, CompileError<'_>>

Compile the function, and emit machine code into a Vec<u8>.

Run the function through all the passes necessary to generate code for the target ISA represented by isa, as well as the final step of emitting machine code into a Vec<u8>. The machine code is not relocated. Instead, any relocations can be obtained from compiled_code().

Performs any optimizations that are enabled, unless optimize() was already invoked.

This function calls compile, taking care to resize mem as needed.

Returns information about the functionโs code and read-only data.

pub fn compile_stencil( &mut self, isa: &dyn TargetIsa ) -> Result<CompiledCodeBase<Stencil>, CodegenError>

Internally compiles the function into a stencil.

Public only for testing and fuzzing purposes.

pub fn optimize(&mut self, isa: &dyn TargetIsa) -> Result<(), CodegenError>

Optimize the function, performing all compilation steps up to but not including machine-code lowering and register allocation.

Public only for testing purposes.

pub fn compile( &mut self, isa: &dyn TargetIsa ) -> Result<&CompiledCodeBase<Final>, CompileError<'_>>

Compile the function.

Run the function through all the passes necessary to generate code for the target ISA represented by isa. This does not include the final step of emitting machine code into a code sink.

Returns information about the functionโs code and read-only data.

pub fn get_code_bb_layout( &self ) -> Option<(Vec<usize, Global>, Vec<(usize, usize), Global>)>

๐Deprecated: use CompiledCode::get_code_bb_layout

If available, return information about the code layout in the final machine code: the offsets (in bytes) of each basic-block start, and all basic-block edges.

pub fn create_unwind_info( &self, isa: &dyn TargetIsa ) -> Result<Option<UnwindInfo>, CodegenError>

๐Deprecated: use CompiledCode::create_unwind_info

Creates unwind information for the function.

Returns None if the function has no unwind information.

pub fn verify<'a, FOI>(&self, fisa: FOI) -> Result<(), VerifierErrors>where FOI: Into<FlagsOrIsa<'a>>,

Run the verifier on the function.

Also check that the dominator tree and control flow graph are consistent with the function.

pub fn verify_if<'a, FOI>(&self, fisa: FOI) -> Result<(), CodegenError>where FOI: Into<FlagsOrIsa<'a>>,

Run the verifier only if the enable_verifier setting is true.

pub fn dce<'a, FOI>(&mut self, fisa: FOI) -> Result<(), CodegenError>where FOI: Into<FlagsOrIsa<'a>>,

Perform dead-code elimination on the function.

pub fn remove_constant_phis<'a, FOI>( &mut self, fisa: FOI ) -> Result<(), CodegenError>where FOI: Into<FlagsOrIsa<'a>>,

Perform constant-phi removal on the function.

pub fn preopt(&mut self, isa: &dyn TargetIsa) -> Result<(), CodegenError>

Perform pre-legalization rewrites on the function.

pub fn canonicalize_nans( &mut self, isa: &dyn TargetIsa ) -> Result<(), CodegenError>

Perform NaN canonicalizing rewrites on the function.

pub fn legalize(&mut self, isa: &dyn TargetIsa) -> Result<(), CodegenError>

Run the legalizer for isa on the function.

pub fn compute_cfg(&mut self)

Compute the control flow graph.

pub fn compute_domtree(&mut self)

Compute dominator tree.

pub fn compute_loop_analysis(&mut self)

Compute the loop analysis.

pub fn flowgraph(&mut self)

Compute the control flow graph and dominator tree.

pub fn simple_gvn<'a, FOI>(&mut self, fisa: FOI) -> Result<(), CodegenError>where FOI: Into<FlagsOrIsa<'a>>,

Perform simple GVN on the function.

pub fn licm(&mut self, isa: &dyn TargetIsa) -> Result<(), CodegenError>

Perform LICM on the function.

pub fn eliminate_unreachable_code<'a, FOI>( &mut self, fisa: FOI ) -> Result<(), CodegenError>where FOI: Into<FlagsOrIsa<'a>>,

Perform unreachable code elimination.

pub fn replace_redundant_loads(&mut self) -> Result<(), CodegenError>

Replace all redundant loads with the known values in memory. These are loads whose values were already loaded by other loads earlier, as well as loads whose values were stored by a store instruction to the same instruction (so-called โstore-to-load forwardingโ).