Struct wasmer_compiler_cranelift::trampoline::FunctionBuilderContext

pub struct FunctionBuilderContext {
    ssa: SSABuilder,
    blocks: SecondaryMap<Block, BlockData>,
    types: SecondaryMap<Variable, Type>,
}

Structure used for translating a series of functions into Cranelift IR.

In order to reduce memory reallocations when compiling multiple functions, FunctionBuilderContext holds various data structures which are cleared between functions, rather than dropped, preserving the underlying allocations.

Fields
ssa: SSABuilderblocks: SecondaryMap<Block, BlockData>types: SecondaryMap<Variable, Type>
Implementations

impl FunctionBuilderContext

pub fn new() -> FunctionBuilderContext

Creates a FunctionBuilderContext structure. The structure is automatically cleared after each FunctionBuilder completes translating a function.