wasmer_compiler_cranelift::translator::code_translator::bounds_checks

Function compute_addr

fn compute_addr(
    pos: &mut FuncCursor<'_>,
    heap: &HeapData,
    addr_ty: Type,
    index: Value,
    offset: u32,
    pcc: Option<AddrPcc>,
) -> Value

Emit code for the native address computation of a Wasm address, without any bounds checks or overflow checks.

It is the caller’s responsibility to ensure that any necessary bounds and overflow checks are emitted, and that the resulting address is never used unless they succeed.