wasmer_compiler_cranelift::translator::code_translator::bounds_checks

Function explicit_check_oob_condition_and_compute_addr

fn explicit_check_oob_condition_and_compute_addr(
    pos: &mut FuncCursor<'_>,
    heap: &HeapData,
    addr_ty: Type,
    index: Value,
    offset: u32,
    spectre_mitigations_enabled: bool,
    oob_condition: Value,
) -> Value

Emit explicit checks on the given out-of-bounds condition for the Wasm address and return the native address.

This function deduplicates explicit bounds checks and Spectre mitigations that inherently also implement bounds checking.