Function wasmer_compiler_cranelift::translator::code_translator::canonicalise_then_jump

fn canonicalise_then_jump(
    builder: &mut FunctionBuilder<'_>, 
    destination: Block, 
    params: (&[Value], &[ValueExtraInfo])
) -> Inst

Generate a jump instruction, but first cast all 128-bit vector values to I8X16 if they don't have that type. This is done in somewhat roundabout way so as to ensure that we almost never have to do any heap allocation.