wasmer_compiler_cranelift::translator::code_translator

Function canonicalise_v128_values

fn canonicalise_v128_values<'a>(
    tmp_canonicalised: &'a mut SmallVec<[BlockArg; 16]>,
    builder: &mut FunctionBuilder<'_>,
    values: &'a [Value],
) -> &'a [BlockArg]

Cast to I8X16, any vector values in values that are of “non-canonical” type (meaning, not I8X16), and return them in a slice. A pre-scan is made to determine whether any casts are actually necessary, and if not, the original slice is returned. Otherwise the cast values are returned in a slice that belongs to the caller-supplied SmallVec.