Function wasmer_compiler_cranelift::translator::code_translator::wasm_param_types

pub fn wasm_param_types(
    params: &[AbiParam],
    is_wasm: impl Fn(usize) -> bool
) -> Vec<Type> 

A helper to extract all the Type listings of each variable in params for only parameters the return true for is_wasm, typically paired with is_wasm_return or is_wasm_parameter.