Macro wasmer_compiler_cranelift::translator::code_translator::unwrap_or_return_unreachable_state

macro_rules! unwrap_or_return_unreachable_state {
    ($state:ident, $value:expr) => { ... };
}

Given a Reachability<T>, unwrap the inner T or, when unreachable, set state.reachable = false and return.

Used in combination with calling prepare_addr and prepare_atomic_addr when we can statically determine that a Wasm access will unconditionally trap.