wasmer_compiler_cranelift::translator

Module func_environ

All the runtime support necessary for the wasm to cranelift translation is formalized by the traits FunctionEnvMutironment.

Enums

GlobalVariable

The value of a WebAssembly global variable.

ReturnMode

How to return from functions.

Traits

FuncEnvironment

Environment affecting the translation of a single WebAssembly function.

TargetEnvironment

Environment affecting the translation of a WebAssembly.