Module wasmer_compiler_cranelift::translator::func_state

WebAssembly module and function translation state.

The ModuleTranslationState struct defined in this module is used to keep track of data about the whole WebAssembly module, such as the decoded type signatures.

The FuncTranslationState struct defined in this module is used to keep track of the WebAssembly value and control stacks during the translation of a single function.

Structs

FuncTranslationState

Contains information passed along during a function’s translation and that records:

Enums

ControlStackFrame

A control stack frame can be an if, a block or a loop, each one having the following fields:

ElseData

Information about the presence of an associated else for an if, or the lack thereof.

Functions

num_wasm_parameters 🔒