wasmer_compiler_cranelift::func_environ

Struct FuncEnvironment

pub struct FuncEnvironment<'module_environment> {

    target_config: TargetFrontendConfig,
    module: &'module_environment ModuleInfo,
    type_stack: Vec<Type>,
    signatures: &'module_environment PrimaryMap<SignatureIndex, Signature>,
    heaps: PrimaryMap<Heap, HeapData>,
    vmctx: Option<GlobalValue>,
    memory32_size_sig: Option<SigRef>,
    table_size_sig: Option<SigRef>,
    memory_grow_sig: Option<SigRef>,
    table_grow_sig: Option<SigRef>,
    table_copy_sig: Option<SigRef>,
    table_init_sig: Option<SigRef>,
    elem_drop_sig: Option<SigRef>,
    memory_copy_sig: Option<SigRef>,
    memory_fill_sig: Option<SigRef>,
    memory_init_sig: Option<SigRef>,
    data_drop_sig: Option<SigRef>,
    table_get_sig: Option<SigRef>,
    table_set_sig: Option<SigRef>,
    func_ref_sig: Option<SigRef>,
    table_fill_sig: Option<SigRef>,
    memory32_atomic_wait32_sig: Option<SigRef>,
    memory32_atomic_wait64_sig: Option<SigRef>,
    memory32_atomic_notify_sig: Option<SigRef>,
    personality2_sig: Option<SigRef>,
    throw_sig: Option<SigRef>,
    alloc_exception_sig: Option<SigRef>,
    read_exception_sig: Option<SigRef>,
    read_exnref_sig: Option<SigRef>,
    exception_type_layouts: HashMap<u32, ExceptionTypeLayout>,
    offsets: VMOffsets,
    memory_styles: &'module_environment PrimaryMap<MemoryIndex, MemoryStyle>,
    tables: SecondaryMap<TableIndex, Option<TableData>>,
    table_styles: &'module_environment PrimaryMap<TableIndex, TableStyle>,

}

The FuncEnvironment implementation for use by the ModuleEnvironment.

Fields
target_config: TargetFrontendConfig

Target-specified configuration.

module: &'module_environment ModuleInfo

The module-level environment which this function-level environment belongs to.

type_stack: Vec<Type>

A stack tracking the type of local variables.

signatures: &'module_environment PrimaryMap<SignatureIndex, Signature>

The module function signatures

heaps: PrimaryMap<Heap, HeapData>

Heaps implementing WebAssembly linear memories.

vmctx: Option<GlobalValue>

The Cranelift global holding the vmctx address.

memory32_size_sig: Option<SigRef>

The external function signature for implementing wasm’s memory.size for locally-defined 32-bit memories.

table_size_sig: Option<SigRef>

The external function signature for implementing wasm’s table.size for locally-defined tables.

memory_grow_sig: Option<SigRef>

The external function signature for implementing wasm’s memory.grow for locally-defined memories.

table_grow_sig: Option<SigRef>

The external function signature for implementing wasm’s table.grow for locally-defined tables.

table_copy_sig: Option<SigRef>

The external function signature for implementing wasm’s table.copy (it’s the same for both local and imported tables).

table_init_sig: Option<SigRef>

The external function signature for implementing wasm’s table.init.

elem_drop_sig: Option<SigRef>

The external function signature for implementing wasm’s elem.drop.

memory_copy_sig: Option<SigRef>

The external function signature for implementing wasm’s memory.copy (it’s the same for both local and imported memories).

memory_fill_sig: Option<SigRef>

The external function signature for implementing wasm’s memory.fill (it’s the same for both local and imported memories).

memory_init_sig: Option<SigRef>

The external function signature for implementing wasm’s memory.init.

data_drop_sig: Option<SigRef>

The external function signature for implementing wasm’s data.drop.

table_get_sig: Option<SigRef>

The external function signature for implementing wasm’s table.get.

table_set_sig: Option<SigRef>

The external function signature for implementing wasm’s table.set.

func_ref_sig: Option<SigRef>

The external function signature for implementing wasm’s func.ref.

table_fill_sig: Option<SigRef>

The external function signature for implementing wasm’s table.fill.

memory32_atomic_wait32_sig: Option<SigRef>

The external function signature for implementing wasm’s memory32.atomic.wait32.

memory32_atomic_wait64_sig: Option<SigRef>

The external function signature for implementing wasm’s memory32.atomic.wait64.

memory32_atomic_notify_sig: Option<SigRef>

The external function signature for implementing wasm’s memory32.atomic.notify.

personality2_sig: Option<SigRef>

Cached signatures for exception helper builtins.

throw_sig: Option<SigRef>alloc_exception_sig: Option<SigRef>read_exception_sig: Option<SigRef>read_exnref_sig: Option<SigRef>exception_type_layouts: HashMap<u32, ExceptionTypeLayout>

Cached payload layouts for exception tags.

offsets: VMOffsets

Offsets to struct fields accessed by JIT code.

memory_styles: &'module_environment PrimaryMap<MemoryIndex, MemoryStyle>

The memory styles

tables: SecondaryMap<TableIndex, Option<TableData>>

Cranelift tables we have created to implement Wasm tables.

table_styles: &'module_environment PrimaryMap<TableIndex, TableStyle>
Implementations

impl<'module_environment> FuncEnvironment<'module_environment>

pub fn new( target_config: TargetFrontendConfig, module: &'module_environment ModuleInfo, signatures: &'module_environment PrimaryMap<SignatureIndex, Signature>, memory_styles: &'module_environment PrimaryMap<MemoryIndex, MemoryStyle>, table_styles: &'module_environment PrimaryMap<TableIndex, TableStyle>, ) -> Self

fn pointer_type(&self) -> Type

fn ensure_table_exists(&mut self, func: &mut Function, index: TableIndex)

fn vmctx(&mut self, func: &mut Function) -> GlobalValue

fn get_table_fill_sig(&mut self, func: &mut Function) -> SigRef

fn get_table_fill_func( &mut self, func: &mut Function, table_index: TableIndex, ) -> (SigRef, usize, VMBuiltinFunctionIndex)

fn get_func_ref_sig(&mut self, func: &mut Function) -> SigRef

fn get_func_ref_func( &mut self, func: &mut Function, function_index: FunctionIndex, ) -> (SigRef, usize, VMBuiltinFunctionIndex)

fn get_table_get_sig(&mut self, func: &mut Function) -> SigRef

fn get_table_get_func( &mut self, func: &mut Function, table_index: TableIndex, ) -> (SigRef, usize, VMBuiltinFunctionIndex)

fn get_table_set_sig(&mut self, func: &mut Function) -> SigRef

fn get_table_set_func( &mut self, func: &mut Function, table_index: TableIndex, ) -> (SigRef, usize, VMBuiltinFunctionIndex)

fn get_table_grow_sig(&mut self, func: &mut Function) -> SigRef

fn get_table_grow_func( &mut self, func: &mut Function, index: TableIndex, ) -> (SigRef, usize, VMBuiltinFunctionIndex)

Return the table.grow function signature to call for the given index, along with the translated index value to pass to it and its index in VMBuiltinFunctionsArray.

fn get_memory_grow_sig(&mut self, func: &mut Function) -> SigRef

fn get_memory_grow_func( &mut self, func: &mut Function, index: MemoryIndex, ) -> (SigRef, usize, VMBuiltinFunctionIndex)

Return the memory.grow function signature to call for the given index, along with the translated index value to pass to it and its index in VMBuiltinFunctionsArray.

fn get_table_size_sig(&mut self, func: &mut Function) -> SigRef

fn get_table_size_func( &mut self, func: &mut Function, index: TableIndex, ) -> (SigRef, usize, VMBuiltinFunctionIndex)

Return the memory.size function signature to call for the given index, along with the translated index value to pass to it and its index in VMBuiltinFunctionsArray.

fn get_memory32_size_sig(&mut self, func: &mut Function) -> SigRef

fn get_memory_size_func( &mut self, func: &mut Function, index: MemoryIndex, ) -> (SigRef, usize, VMBuiltinFunctionIndex)

Return the memory.size function signature to call for the given index, along with the translated index value to pass to it and its index in VMBuiltinFunctionsArray.

fn get_table_copy_sig(&mut self, func: &mut Function) -> SigRef

fn get_table_copy_func( &mut self, func: &mut Function, dst_table_index: TableIndex, src_table_index: TableIndex, ) -> (SigRef, usize, usize, VMBuiltinFunctionIndex)

fn get_table_init_sig(&mut self, func: &mut Function) -> SigRef

fn get_table_init_func( &mut self, func: &mut Function, table_index: TableIndex, ) -> (SigRef, usize, VMBuiltinFunctionIndex)

fn get_elem_drop_sig(&mut self, func: &mut Function) -> SigRef

fn get_elem_drop_func( &mut self, func: &mut Function, ) -> (SigRef, VMBuiltinFunctionIndex)

fn get_memory_copy_sig(&mut self, func: &mut Function) -> SigRef

fn get_memory_copy_func( &mut self, func: &mut Function, memory_index: MemoryIndex, ) -> (SigRef, usize, VMBuiltinFunctionIndex)

fn get_memory_fill_sig(&mut self, func: &mut Function) -> SigRef

fn get_memory_fill_func( &mut self, func: &mut Function, memory_index: MemoryIndex, ) -> (SigRef, usize, VMBuiltinFunctionIndex)

fn get_memory_init_sig(&mut self, func: &mut Function) -> SigRef

fn get_memory_init_func( &mut self, func: &mut Function, ) -> (SigRef, VMBuiltinFunctionIndex)

fn get_data_drop_sig(&mut self, func: &mut Function) -> SigRef

fn get_data_drop_func( &mut self, func: &mut Function, ) -> (SigRef, VMBuiltinFunctionIndex)

fn get_memory32_atomic_wait32_sig(&mut self, func: &mut Function) -> SigRef

fn get_memory_atomic_wait32_func( &mut self, func: &mut Function, index: MemoryIndex, ) -> (SigRef, usize, VMBuiltinFunctionIndex)

Return the memory.atomic.wait32 function signature to call for the given index, along with the translated index value to pass to it and its index in VMBuiltinFunctionsArray.

fn get_memory32_atomic_wait64_sig(&mut self, func: &mut Function) -> SigRef

fn get_memory_atomic_wait64_func( &mut self, func: &mut Function, index: MemoryIndex, ) -> (SigRef, usize, VMBuiltinFunctionIndex)

Return the memory.atomic.wait64 function signature to call for the given index, along with the translated index value to pass to it and its index in VMBuiltinFunctionsArray.

fn get_memory32_atomic_notify_sig(&mut self, func: &mut Function) -> SigRef

fn get_memory_atomic_notify_func( &mut self, func: &mut Function, index: MemoryIndex, ) -> (SigRef, usize, VMBuiltinFunctionIndex)

Return the memory.atomic.notify function signature to call for the given index, along with the translated index value to pass to it and its index in VMBuiltinFunctionsArray.

fn get_personality2_func( &mut self, func: &mut Function, ) -> (SigRef, VMBuiltinFunctionIndex)

fn get_throw_func( &mut self, func: &mut Function, ) -> (SigRef, VMBuiltinFunctionIndex)

fn get_alloc_exception_func( &mut self, func: &mut Function, ) -> (SigRef, VMBuiltinFunctionIndex)

fn get_read_exnref_func( &mut self, func: &mut Function, ) -> (SigRef, VMBuiltinFunctionIndex)

fn get_read_exception_func( &mut self, func: &mut Function, ) -> (SigRef, VMBuiltinFunctionIndex)

fn exception_type_layout( &mut self, tag_index: TagIndex, ) -> WasmResult<&ExceptionTypeLayout>

fn compute_exception_type_layout( &self, tag_index: TagIndex, ) -> WasmResult<ExceptionTypeLayout>

fn map_wasmer_type_to_ir(&self, ty: WasmerType) -> WasmResult<Type>

fn call_with_handlers( &mut self, builder: &mut FunctionBuilder<'_>, callee: FuncRef, args: &[Value], context: Option<Value>, landing_pad: Option<LandingPad>, unreachable_on_return: bool, ) -> SmallVec<[Value; 4]>

fn call_indirect_with_handlers( &mut self, builder: &mut FunctionBuilder<'_>, sig: SigRef, func_addr: Value, args: &[Value], context: Option<Value>, landing_pad: Option<LandingPad>, unreachable_on_return: bool, ) -> SmallVec<[Value; 4]>

fn translate_load_builtin_function_address( &mut self, pos: &mut FuncCursor<'_>, callee_func_idx: VMBuiltinFunctionIndex, ) -> (Value, Value)

Translates load of builtin function and returns a pair of values vmctx and address of the loaded function.

fn get_or_init_funcref_table_elem( &mut self, builder: &mut FunctionBuilder<'_>, table_index: TableIndex, index: Value, ) -> Value