wasmer_compiler_cranelift
Module trampoline

Re-exports

pub use self::dynamic_function::make_trampoline_dynamic_function;

pub use self::function_call::make_trampoline_function_call;

Modules

dynamic_function 🔒

A trampoline generator for calling dynamic host functions from Wasm.

function_call 🔒

A trampoline generator for calling Wasm functions easily.

Structs

FunctionBuilderContext

Structure used for translating a series of functions into Cranelift IR.