Module wasmer_compiler_cranelift::trampoline

Re-exports

pub use self::dynamic_function::make_trampoline_dynamic_function;

pub use self::function_call::make_trampoline_function_call;

Modules

binemit

dynamic_function

A trampoline generator for calling dynamic host functions from Wasm.

function_call

A trampoline generator for calling Wasm functions easily.

Structs

Context

Persistent data structures and compilation pipeline.

FunctionBuilder

Temporary object used to build a single Cranelift IR Function.

FunctionBuilderContext

Structure used for translating a series of functions into Cranelift IR.

Functions

pretty_error

Pretty-print a Cranelift error.