Module wasmer_compiler_cranelift::translator

Tools for translating wasm function bytecode to Cranelift IR.

Re-exports

pub use self::func_environ::FuncEnvironment;

pub use self::func_environ::GlobalVariable;

pub use self::func_environ::ReturnMode;

pub use self::func_environ::TargetEnvironment;

pub use self::func_state::FuncTranslationState;

pub use self::func_translator::FuncTranslator;

pub use self::translation_utils::get_vmctx_value_label;

pub use self::translation_utils::irlibcall_to_libcall;

pub use self::translation_utils::irreloc_to_relocationkind;

pub use self::translation_utils::signature_to_cranelift_ir;

pub use self::translation_utils::transform_jump_table;

pub use self::translation_utils::type_to_irtype;

Modules

code_translator
This module contains the bulk of the interesting code performing the translation between WebAssembly bytecode and Cranelift IR.

func_environ
All the runtime support necessary for the wasm to cranelift translation is formalized by the traits FunctionEnvironment.

func_state
WebAssembly module and function translation state.

func_translator
Standalone WebAssembly to Cranelift IR translator.

translation_utils
Helper functions and structures for the translation.

unwind
A Compilation contains the compiled function bodies for a WebAssembly module.