wasmer_compiler_cranelift::translator::code_translator::bounds_checks
Function bounds_check_and_compute_addr

pub fn bounds_check_and_compute_addr<Env>(
    builder: &mut FunctionBuilder<'_>,
    env: &mut Env,
    heap: &HeapData,
    index: Value,
    offset: u32,
    access_size: u8,
) -> WasmResult<Reachability<Value>>
where
    Env: FuncEnvironment + ?Sized,

Helper used to emit bounds checks (as necessary) and compute the native address of a heap access.

Returns the ir::Value holding the native address of the heap access, or None if the heap access will unconditionally trap.