Trait wasmer_compiler_cranelift::translator::func_environ::FuncEnvironment

pub trait FuncEnvironment: TargetEnvironment {

    fn make_global(
        &mut self, 
        func: &mut Function, 
        index: GlobalIndex
    ) -> WasmResult<GlobalVariable>;

    fn make_heap(
        &mut self, 
        func: &mut Function, 
        index: MemoryIndex
    ) -> WasmResult<Heap>;

    fn make_table(
        &mut self, 
        func: &mut Function, 
        index: TableIndex
    ) -> WasmResult<Table>;

    fn make_indirect_sig(
        &mut self, 
        func: &mut Function, 
        index: SignatureIndex
    ) -> WasmResult<SigRef>;

    fn make_direct_func(
        &mut self, 
        func: &mut Function, 
        index: FunctionIndex
    ) -> WasmResult<FuncRef>;

    fn translate_call_indirect(
        &mut self, 
        pos: FuncCursor<'_>, 
        table_index: TableIndex, 
        table: Table, 
        sig_index: SignatureIndex, 
        sig_ref: SigRef, 
        callee: Value, 
        call_args: &[Value]
    ) -> WasmResult<Inst>;

    fn translate_memory_grow(
        &mut self, 
        pos: FuncCursor<'_>, 
        index: MemoryIndex, 
        heap: Heap, 
        val: Value
    ) -> WasmResult<Value>;

    fn translate_memory_size(
        &mut self, 
        pos: FuncCursor<'_>, 
        index: MemoryIndex, 
        heap: Heap
    ) -> WasmResult<Value>;

    fn translate_memory_copy(
        &mut self, 
        pos: FuncCursor<'_>, 
        src_index: MemoryIndex, 
        src_heap: Heap, 
        dst_index: MemoryIndex, 
        dst_heap: Heap, 
        dst: Value, 
        src: Value, 
        len: Value
    ) -> WasmResult<()>;

    fn translate_memory_fill(
        &mut self, 
        pos: FuncCursor<'_>, 
        index: MemoryIndex, 
        heap: Heap, 
        dst: Value, 
        val: Value, 
        len: Value
    ) -> WasmResult<()>;

    fn translate_memory_init(
        &mut self, 
        pos: FuncCursor<'_>, 
        index: MemoryIndex, 
        heap: Heap, 
        seg_index: u32, 
        dst: Value, 
        src: Value, 
        len: Value
    ) -> WasmResult<()>;

    fn translate_data_drop(
        &mut self, 
        pos: FuncCursor<'_>, 
        seg_index: u32
    ) -> WasmResult<()>;

    fn translate_table_size(
        &mut self, 
        pos: FuncCursor<'_>, 
        index: TableIndex, 
        table: Table
    ) -> WasmResult<Value>;

    fn translate_table_grow(
        &mut self, 
        pos: FuncCursor<'_>, 
        table_index: TableIndex, 
        table: Table, 
        delta: Value, 
        init_value: Value
    ) -> WasmResult<Value>;

    fn translate_table_get(
        &mut self, 
        builder: &mut FunctionBuilder<'_>, 
        table_index: TableIndex, 
        table: Table, 
        index: Value
    ) -> WasmResult<Value>;

    fn translate_table_set(
        &mut self, 
        builder: &mut FunctionBuilder<'_>, 
        table_index: TableIndex, 
        table: Table, 
        value: Value, 
        index: Value
    ) -> WasmResult<()>;

    fn translate_table_copy(
        &mut self, 
        pos: FuncCursor<'_>, 
        dst_table_index: TableIndex, 
        dst_table: Table, 
        src_table_index: TableIndex, 
        src_table: Table, 
        dst: Value, 
        src: Value, 
        len: Value
    ) -> WasmResult<()>;

    fn translate_table_fill(
        &mut self, 
        pos: FuncCursor<'_>, 
        table_index: TableIndex, 
        dst: Value, 
        val: Value, 
        len: Value
    ) -> WasmResult<()>;

    fn translate_externref_inc(
        &mut self, 
        pos: FuncCursor<'_>, 
        externref: Value
    ) -> WasmResult<()>;

    fn translate_externref_dec(
        &mut self, 
        pos: FuncCursor<'_>, 
        externref: Value
    ) -> WasmResult<()>;

    fn translate_table_init(
        &mut self, 
        pos: FuncCursor<'_>, 
        seg_index: u32, 
        table_index: TableIndex, 
        table: Table, 
        dst: Value, 
        src: Value, 
        len: Value
    ) -> WasmResult<()>;

    fn translate_elem_drop(
        &mut self, 
        pos: FuncCursor<'_>, 
        seg_index: u32
    ) -> WasmResult<()>;

    fn translate_ref_null(
        &mut self, 
        pos: FuncCursor<'_>, 
        ty: Type
    ) -> WasmResult<Value>;

    fn translate_ref_func(
        &mut self, 
        pos: FuncCursor<'_>, 
        func_index: FunctionIndex
    ) -> WasmResult<Value>;

    fn translate_custom_global_get(
        &mut self, 
        pos: FuncCursor<'_>, 
        global_index: GlobalIndex
    ) -> WasmResult<Value>;

    fn translate_custom_global_set(
        &mut self, 
        pos: FuncCursor<'_>, 
        global_index: GlobalIndex, 
        val: Value
    ) -> WasmResult<()>;

    fn translate_atomic_wait(
        &mut self, 
        pos: FuncCursor<'_>, 
        index: MemoryIndex, 
        heap: Heap, 
        addr: Value, 
        expected: Value, 
        timeout: Value
    ) -> WasmResult<Value>;

    fn translate_atomic_notify(
        &mut self, 
        pos: FuncCursor<'_>, 
        index: MemoryIndex, 
        heap: Heap, 
        addr: Value, 
        count: Value
    ) -> WasmResult<Value>;

    fn get_global_type(&self, global_index: GlobalIndex) -> Option<WasmerType>;

    fn push_local_decl_on_stack(&mut self, ty: WasmerType);

    fn push_params_on_stack(&mut self, function_index: LocalFunctionIndex);

    fn get_local_type(&self, local_index: u32) -> Option<WasmerType>;

    fn get_local_types(&self) -> &[WasmerType];

    fn get_function_type(
        &self, 
        function_index: FunctionIndex
    ) -> Option<&FunctionType>;

    fn get_function_sig(
        &self, 
        sig_index: SignatureIndex
    ) -> Option<&FunctionType>;

    fn translate_drop_locals(
        &mut self, 
        builder: &mut FunctionBuilder<'_>
    ) -> WasmResult<()>;

    fn is_wasm_parameter(&self, signature: &Signature, index: usize) -> bool { ... }

    fn is_wasm_return(&self, signature: &Signature, index: usize) -> bool { ... }

    fn return_mode(&self) -> ReturnMode { ... }

    fn translate_call(
        &mut self, 
        pos: FuncCursor<'_>, 
        _callee_index: FunctionIndex, 
        callee: FuncRef, 
        call_args: &[Value]
    ) -> WasmResult<Inst> { ... }

    fn translate_ref_is_null(
        &mut self, 
        pos: FuncCursor<'_>, 
        value: Value
    ) -> WasmResult<Value> { ... }

    fn translate_loop_header(&mut self, _pos: FuncCursor<'_>) -> WasmResult<()> { ... }

    fn before_translate_operator(
        &mut self, 
        _op: &Operator<'_>, 
        _builder: &mut FunctionBuilder<'_>, 
        _state: &FuncTranslationState
    ) -> WasmResult<()> { ... }

    fn after_translate_operator(
        &mut self, 
        _op: &Operator<'_>, 
        _builder: &mut FunctionBuilder<'_>, 
        _state: &FuncTranslationState
    ) -> WasmResult<()> { ... }

}

Environment affecting the translation of a single WebAssembly function.

A FuncEnvironment trait object is required to translate a WebAssembly function to Cranelift IR. The function environment provides information about the WebAssembly module as well as the runtime environment.

Required methods

fn make_global(
&mut self,
func: &mut Function,
index: GlobalIndex
) -> WasmResult<GlobalVariable>

Set up the necessary preamble definitions in func to access the global variable identified by index.

The index space covers both imported globals and globals defined by the module.

Return the global variable reference that should be used to access the global and the WebAssembly type of the global.

fn make_heap(
&mut self,
func: &mut Function,
index: MemoryIndex
) -> WasmResult<Heap>

Set up the necessary preamble definitions in func to access the linear memory identified by index.

The index space covers both imported and locally declared memories.

fn make_table(
&mut self,
func: &mut Function,
index: TableIndex
) -> WasmResult<Table>

Set up the necessary preamble definitions in func to access the table identified by index.

The index space covers both imported and locally declared tables.

fn make_indirect_sig(
&mut self,
func: &mut Function,
index: SignatureIndex
) -> WasmResult<SigRef>

Set up a signature definition in the preamble of func that can be used for an indirect call with signature index.

The signature may contain additional arguments needed for an indirect call, but the arguments marked as ArgumentPurpose::Normal must correspond to the WebAssembly signature arguments.

The signature will only be used for indirect calls, even if the module has direct function calls with the same WebAssembly type.

fn make_direct_func(
&mut self,
func: &mut Function,
index: FunctionIndex
) -> WasmResult<FuncRef>

Set up an external function definition in the preamble of func that can be used to directly call the function index.

The index space covers both imported functions and functions defined in the current module.

The function’s signature may contain additional arguments needed for a direct call, but the arguments marked as ArgumentPurpose::Normal must correspond to the WebAssembly signature arguments.

The function’s signature will only be used for direct calls, even if the module has indirect calls with the same WebAssembly type.

fn translate_call_indirect(
&mut self,
pos: FuncCursor<'_>,
table_index: TableIndex,
table: Table,
sig_index: SignatureIndex,
sig_ref: SigRef,
callee: Value,
call_args: &[Value]
) -> WasmResult<Inst>

Translate a call_indirect WebAssembly instruction at pos.

Insert instructions at pos for an indirect call to the function callee in the table table_index with WebAssembly signature sig_index. The callee value will have type i32.

The signature sig_ref was previously created by make_indirect_sig().

Return the call instruction whose results are the WebAssembly return values.

fn translate_memory_grow(
&mut self,
pos: FuncCursor<'_>,
index: MemoryIndex,
heap: Heap,
val: Value
) -> WasmResult<Value>

Translate a memory.grow WebAssembly instruction.

The index provided identifies the linear memory to grow, and heap is the heap reference returned by make_heap for the same index.

The val value is the requested memory size in pages.

Returns the old size (in pages) of the memory.

fn translate_memory_size(
&mut self,
pos: FuncCursor<'_>,
index: MemoryIndex,
heap: Heap
) -> WasmResult<Value>

Translates a memory.size WebAssembly instruction.

The index provided identifies the linear memory to query, and heap is the heap reference returned by make_heap for the same index.

Returns the size in pages of the memory.

fn translate_memory_copy(
&mut self,
pos: FuncCursor<'_>,
src_index: MemoryIndex,
src_heap: Heap,
dst_index: MemoryIndex,
dst_heap: Heap,
dst: Value,
src: Value,
len: Value
) -> WasmResult<()>

Translate a memory.copy WebAssembly instruction.

The index provided identifies the linear memory to query, and heap is the heap reference returned by make_heap for the same index.

fn translate_memory_fill(
&mut self,
pos: FuncCursor<'_>,
index: MemoryIndex,
heap: Heap,
dst: Value,
val: Value,
len: Value
) -> WasmResult<()>

Translate a memory.fill WebAssembly instruction.

The index provided identifies the linear memory to query, and heap is the heap reference returned by make_heap for the same index.

fn translate_memory_init(
&mut self,
pos: FuncCursor<'_>,
index: MemoryIndex,
heap: Heap,
seg_index: u32,
dst: Value,
src: Value,
len: Value
) -> WasmResult<()>

Translate a memory.init WebAssembly instruction.

The index provided identifies the linear memory to query, and heap is the heap reference returned by make_heap for the same index. seg_index is the index of the segment to copy from.

fn translate_data_drop(
&mut self,
pos: FuncCursor<'_>,
seg_index: u32
) -> WasmResult<()>

Translate a data.drop WebAssembly instruction.

fn translate_table_size(
&mut self,
pos: FuncCursor<'_>,
index: TableIndex,
table: Table
) -> WasmResult<Value>

Translate a table.size WebAssembly instruction.

fn translate_table_grow(
&mut self,
pos: FuncCursor<'_>,
table_index: TableIndex,
table: Table,
delta: Value,
init_value: Value
) -> WasmResult<Value>

Translate a table.grow WebAssembly instruction.

fn translate_table_get(
&mut self,
builder: &mut FunctionBuilder<'_>,
table_index: TableIndex,
table: Table,
index: Value
) -> WasmResult<Value>

Translate a table.get WebAssembly instruction.

fn translate_table_set(
&mut self,
builder: &mut FunctionBuilder<'_>,
table_index: TableIndex,
table: Table,
value: Value,
index: Value
) -> WasmResult<()>

Translate a table.set WebAssembly instruction.

fn translate_table_copy(
&mut self,
pos: FuncCursor<'_>,
dst_table_index: TableIndex,
dst_table: Table,
src_table_index: TableIndex,
src_table: Table,
dst: Value,
src: Value,
len: Value
) -> WasmResult<()>

Translate a table.copy WebAssembly instruction.

fn translate_table_fill(
&mut self,
pos: FuncCursor<'_>,
table_index: TableIndex,
dst: Value,
val: Value,
len: Value
) -> WasmResult<()>

Translate a table.fill WebAssembly instruction.

fn translate_externref_inc(
&mut self,
pos: FuncCursor<'_>,
externref: Value
) -> WasmResult<()>

Translates an externref ref count increment.

fn translate_externref_dec(
&mut self,
pos: FuncCursor<'_>,
externref: Value
) -> WasmResult<()>

Translates an externref ref count decrement.

fn translate_table_init(
&mut self,
pos: FuncCursor<'_>,
seg_index: u32,
table_index: TableIndex,
table: Table,
dst: Value,
src: Value,
len: Value
) -> WasmResult<()>

Translate a table.init WebAssembly instruction.

fn translate_elem_drop(
&mut self,
pos: FuncCursor<'_>,
seg_index: u32
) -> WasmResult<()>

Translate a elem.drop WebAssembly instruction.

fn translate_ref_null(
&mut self,
pos: FuncCursor<'_>,
ty: Type
) -> WasmResult<Value>

Translate a ref.null T WebAssembly instruction.

By default, translates into a null reference type.

Override this if you don’t use Cranelift reference types for all Wasm reference types (e.g. you use a raw pointer for funcrefs) or if the null sentinel is not a null reference type pointer for your type. If you override this method, then you should also override translate_ref_is_null as well.

fn translate_ref_func(
&mut self,
pos: FuncCursor<'_>,
func_index: FunctionIndex
) -> WasmResult<Value>

Translate a ref.func WebAssembly instruction.

fn translate_custom_global_get(
&mut self,
pos: FuncCursor<'_>,
global_index: GlobalIndex
) -> WasmResult<Value>

Translate a global.get WebAssembly instruction at pos for a global that is custom.

fn translate_custom_global_set(
&mut self,
pos: FuncCursor<'_>,
global_index: GlobalIndex,
val: Value
) -> WasmResult<()>

Translate a global.set WebAssembly instruction at pos for a global that is custom.

fn translate_atomic_wait(
&mut self,
pos: FuncCursor<'_>,
index: MemoryIndex,
heap: Heap,
addr: Value,
expected: Value,
timeout: Value
) -> WasmResult<Value>

Translate an i32.atomic.wait or i64.atomic.wait WebAssembly instruction. The index provided identifies the linear memory containing the value to wait on, and heap is the heap reference returned by make_heap for the same index. Whether the waited-on value is 32- or 64-bit can be determined by examining the type of expected, which must be only I32 or I64.

Returns an i32, which is negative if the helper call failed.

fn translate_atomic_notify(
&mut self,
pos: FuncCursor<'_>,
index: MemoryIndex,
heap: Heap,
addr: Value,
count: Value
) -> WasmResult<Value>

Translate an atomic.notify WebAssembly instruction. The index provided identifies the linear memory containing the value to wait on, and heap is the heap reference returned by make_heap for the same index.

Returns an i64, which is negative if the helper call failed.

fn get_global_type(&self, global_index: GlobalIndex) -> Option<WasmerType>

Get the type of the global at the given index.

fn push_local_decl_on_stack(&mut self, ty: WasmerType)

Push a local declaration on to the stack to track the type of locals.

fn push_params_on_stack(&mut self, function_index: LocalFunctionIndex)

Push locals for a the params of a function on to the stack.

fn get_local_type(&self, local_index: u32) -> Option<WasmerType>

Get the type of the local at the given index.

fn get_local_types(&self) -> &[WasmerType]

Get the types of all the current locals.

fn get_function_type(
&self,
function_index: FunctionIndex
) -> Option<&FunctionType>

Get the type of the local at the given index.

fn get_function_sig(&self, sig_index: SignatureIndex) -> Option<&FunctionType>

Get the type of a function with the given signature index.

fn translate_drop_locals(
&mut self,
builder: &mut FunctionBuilder<'_>
) -> WasmResult<()>

Drops all locals that need to be dropped. Useful for returning from functions.

Provided methods

fn is_wasm_parameter(&self, signature: &Signature, index: usize) -> bool

Is the given parameter of the given function a wasm-level parameter, as opposed to a hidden parameter added for use by the implementation?

fn is_wasm_return(&self, signature: &Signature, index: usize) -> bool

Is the given return of the given function a wasm-level parameter, as opposed to a hidden parameter added for use by the implementation?

fn return_mode(&self) -> ReturnMode

Should the code be structured to use a single fallthrough_return instruction at the end of the function body, rather than return instructions as needed? This is used by VMs to append custom epilogues.

fn translate_call(
&mut self,
pos: FuncCursor<'_>,
_callee_index: FunctionIndex,
callee: FuncRef,
call_args: &[Value]
) -> WasmResult<Inst>

Translate a call WebAssembly instruction at pos.

Insert instructions at pos for a direct call to the function callee_index.

The function reference callee was previously created by make_direct_func().

Return the call instruction whose results are the WebAssembly return values.

fn translate_ref_is_null(
&mut self,
pos: FuncCursor<'_>,
value: Value
) -> WasmResult<Value>

Translate a ref.is_null WebAssembly instruction.

By default, assumes that value is a Cranelift reference type, and that a null Cranelift reference type is the null value for all Wasm reference types.

If you override this method, you probably also want to override translate_ref_null as well.

fn translate_loop_header(&mut self, _pos: FuncCursor<'_>) -> WasmResult<()>

Emit code at the beginning of every wasm loop.

This can be used to insert explicit interrupt or safepoint checking at the beginnings of loops.

fn before_translate_operator(
&mut self,
_op: &Operator<'_>,
_builder: &mut FunctionBuilder<'_>,
_state: &FuncTranslationState
) -> WasmResult<()>

Optional callback for the FunctionEnvironment performing this translation to maintain internal state or prepare custom state for the operator to translate

fn after_translate_operator(
&mut self,
_op: &Operator<'_>,
_builder: &mut FunctionBuilder<'_>,
_state: &FuncTranslationState
) -> WasmResult<()>

Optional callback for the FunctionEnvironment performing this translation to maintain internal state or finalize custom state for the operator that was translated