Trait wasmer_compiler_cranelift::translator::func_environ::TargetEnvironment

pub trait TargetEnvironment {
    fn target_config(&self) -> TargetFrontendConfig;

    fn pointer_type(&self) -> Type { ... }

    fn pointer_bytes(&self) -> u8 { ... }

    fn reference_type(&self) -> Type { ... }
}

Environment affecting the translation of a WebAssembly.

Required methods

fn target_config(&self) -> TargetFrontendConfig

Get the information needed to produce Cranelift IR for the given target.

Provided methods

fn pointer_type(&self) -> Type

Get the Cranelift integer type to use for native pointers.

This returns I64 for 64-bit architectures and I32 for 32-bit architectures.

fn pointer_bytes(&self) -> u8

Get the size of a native pointer, in bytes.

fn reference_type(&self) -> Type

Get the Cranelift reference type to use for native references.

This returns R64 for 64-bit architectures and R32 for 32-bit architectures.