Module wasmer_compiler_cranelift::translator::func_translator

Standalone WebAssembly to Cranelift IR translator.

This module defines the FuncTranslator type which can translate a single WebAssembly function to Cranelift IR guided by a FuncEnvironment which provides information about the WebAssembly module and the runtime environment.

Structs

FuncTranslator

WebAssembly to Cranelift IR function translator.

Functions

cur_srcloc

Get the current source location from a reader.

declare_locals

Declare count local variables of the same type, starting from next_local.

declare_wasm_parameters

Declare local variables for the signature parameters that correspond to WebAssembly locals.

parse_function_body

Parse the function body in reader.

parse_local_decls

Parse the local variable declarations that precede the function body.