wasmer_compiler_cranelift::translator
Module translation_utils

Helper functions and structures for the translation.

Functions

block_with_params

Create a Block with the given Wasm parameters.

f32_translation

Turns a wasmparser f32 into a Cranelift one.

f64_translation

Turns a wasmparser f64 into a Cranelift one.

get_vmctx_value_label

Special VMContext value label. It is tracked as 0xffff_fffe label.

irlibcall_to_libcall

Transform Cranelift LibCall into runtime LibCall

irreloc_to_relocationkind

Transform Cranelift Reloc to compiler Relocation

reference_type

Helper function translating wasmparser types to Cranelift types when possible.

signature_to_cranelift_ir

Helper function translate a Function signature into Cranelift Ir

type_to_irtype

Helper function translating wasmparser types to Cranelift types when possible.